update: pengembangan kompetensi

+++ app/Views/pengembangan_kompetensi/main.php
@@ -1,12 +1,16 @@
 <?php
-$is_access = true;
-// if (array_intersect(userSession('id_roles'), [1, 17, 4])) {
-//     $is_access = true;
-// }
+$is_access = false;
+if (array_intersect(userSession('id_roles'), [4])) {
+    $is_access = true;
+}
+
+$get_dosen = $_GET['dosen'] ?? '';
 ?>
 
 <script src="<?= base_url() ?>assets/js/jquery.min.js"></script>
 <link rel="stylesheet" href="<?= base_url() ?>assets/modules/datatables/css/dataTables.dataTables.min.css">
+<link rel="stylesheet" href="<?= base_url() ?>assets/modules/dselect/dselect.min.css">
+<script src="<?= base_url() ?>assets/modules/dselect/dselect.min.js"></script>
 
 <section class="container-fluid">
     <div class="row">
@@ -14,66 +18,77 @@ $is_access = true;
             <h4 class="my-4"><?= isset($title) ? $title : '' ?></h4>
         </div>
     </div>
+    <?php if (in_array(8, userSession('id_roles'))) : ?>
+    <div class="row">
+        <div class="col-12">
+            <div class="alert alert-primary">
+                Sebagai Kaprodi, Anda dapat melihat semua data dosen <span class="fw-500"><?= userSession('nama_program_studi') ?></span>
+            </div>
+        </div>
+    </div>
+    <?php endif; ?>
     <div class="row">
         <div class="col-12">
             <div class="card p-3">
-                <?php if ($is_access) : ?>
                 <div class="row g-3 mb-3">
-                    <div class="col-12 col-md-6 col-lg-5 col-xl-4">
-                        <!--  -->
-                    </div>
-                    <div class="col-12 col-md-6 col-lg-7 col-xl-8 d-flex justify-content-end align-items-end">
-                        <button class="btn btn-primary" data-bs-toggle="modal" data-bs-target="#new">
-                            <i class="fa-solid fa-plus fa-sm"></i> New
-                        </button>
-                        <div class="modal fade" id="new" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1">
-                            <div class="modal-dialog">
-                                <div class="modal-content">
-                                    <div class="modal-header">
-                                        <h1 class="modal-title fs-5">Add <?= isset($title) ? $title : '' ?></h1>
-                                        <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
-                                    </div>
-                                    <form id="form">
-                                        <div class="modal-body">
-                                            <div class="mb-3">
-                                                <div class="mb-3">
-                                                    <label for="judul" class="form-label">Judul</label>
-                                                    <input type="text" class="form-control" id="judul" name="judul" placeholder="Masukkan judul">
-                                                    <div class="invalid-feedback" id="invalid_judul"></div>
-                                                </div>
-                                                <div class="mb-3">
-                                                    <label for="tautan" class="form-label">Tautan</label>
-                                                    <input type="text" class="form-control" id="tautan" name="tautan" placeholder="Masukkan tautan">
-                                                    <div class="invalid-feedback" id="invalid_tautan"></div>
-                                                </div>
-                                                <div class="mb-3">
-                                                    <label for="dokumen" class="form-label">Dokumen</label>
-                                                    <input type="file" class="form-control" id="dokumen" name="dokumen" accept="application/pdf">
-                                                    <div class="form-text">
-                                                        Maksimal 1 mb, pdf
-                                                    </div>
-                                                    <div class="invalid-feedback" id="invalid_dokumen"></div>
-                                                </div>
-                                            </div>
-                                        </div>
-                                        <div class="modal-footer">
-                                            <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Batal</button>
-                                            <button type="submit" class="btn btn-primary float-end">Tambahkan</button>
-                                        </div>
-                                    </form>
-                                    <script>
-                                    dom('#form').addEventListener('submit', function(event) {
-                                        event.preventDefault();
-                                        const endpoint = '<?= $base_api ?>create';
-                                        submitData(dom('#form'), endpoint);
-                                    });
-                                    </script>
+                    <div class="col-12 col-lg-10 col-xl-11">
+                        <?php if (array_intersect(userSession('id_roles'), [1, 17, 7, 8])) : ?>
+                        <form action="" method="get">
+                            <div class="row gx-2 gy-3">
+                                <div class="col-12 col-md-5 col-lg-5 col-xl-4">
+                                    <label for="dosen" class="form-label">Dosen</label>
+                                    <select id="dosen" name="dosen">
+                                        <option value="">Pilih</option>
+                                        <?php
+                                        $dosen = [];
+                                        if (in_array(userSession('id'), [1, 17, 7])) {
+                                            $dosen = model('Users')->where('id_role', 4)->findAll();
+                                        } elseif (in_array(8, userSession('id_roles'))) {
+                                            $dosen = model('Users')->where('id_program_studi', userSession('id_program_studi'))->findAll();
+                                        }
+                                        foreach ($dosen as $v) :
+                                            $selected = ($v['id'] == $get_dosen) ? 'selected' : '';
+                                        ?>
+                                        <option value="<?= $v['id'] ?>" <?= $selected ?>>
+                                            <?= $v['nomor_identitas'] ?> - <?= $v['nama'] ?> - <?= $v['nama_program_studi'] ?>
+                                        </option>
+                                        <?php endforeach; ?>
+                                    </select>
                                 </div>
+                                <div class="col-12 col-md-2 col-lg-2 col-xl-2 d-flex justify-content-start align-items-end">
+                                    <button type="submit" class="btn btn-primary me-2 w-100" title="Filter">
+                                        <i class="fa-solid fa-filter"></i>
+                                        <span class="ms-1 d-md-none">Filter</span>
+                                    </button>
+                                    <a href="<?= $base_route ?>" class="btn btn-outline-danger w-100" title="Reset">
+                                        <i class="fa-solid fa-filter-circle-xmark"></i>
+                                        <span class="ms-1 d-md-none">Reset</span>
+                                    </a>
+                                </div>
+                                <?php if (in_array(8, userSession('id_roles'))) : ?>
+                                <div class="col-6 col-md-5 col-lg-4 col-xl-3 d-flex justify-content-bottom align-items-end">
+                                    <?php if ($get_dosen != userSession('id')) : ?>
+                                    <a href="<?= $base_route ?>?dosen=<?= userSession('id') ?>" class="btn btn-primary">Lihat Data Saya</a>
+                                    <?php else : ?>
+                                    <a href="<?= $base_route ?>" class="btn btn-primary">Lihat Semua Data Dosen</a>
+                                    <?php endif; ?>
+                                </div>
+                                <?php endif; ?>
                             </div>
-                        </div>
+                        </form>
+                        <script>
+                        dselect(dom('#dosen'), { search: true, clearable: true });
+                        </script>
+                        <?php endif; ?>
+                    </div>
+                    <div class="col-12 col-lg-2 col-xl-1 d-flex justify-content-end align-items-end">
+                        <?php if ($is_access) : ?>
+                        <a href="<?= $base_route ?>new" class="btn btn-primary">
+                            <i class="fa-solid fa-plus fa-sm"></i> New
+                        </a>
+                        <?php endif; ?>
                     </div>
                 </div>
-                <?php endif; ?>
                 <table class="table table-striped table-hover table-bordered text-nowrap" id="myTable">
                     <thead class="bg-primary-subtle">
                         <tr>
@@ -136,66 +151,19 @@ document.addEventListener('DOMContentLoaded', function() {
 
 <?php if ($is_access) : ?>
 function renderOpsi(data) {
+    let endpoint_edit_data = `<?= $base_route ?>edit/${data.id}`;
     let endpoint_hapus_data = `<?= $base_api ?>delete/${data.id}`;
-    let html = `
-    <a class="me-2" title="Edit" data-bs-toggle="modal" data-bs-target="#edit${data.id}">
-        <i class="fa-regular fa-pen-to-square fa-lg"></i>
-    </a>
-    <div class="modal fade" id="edit${data.id}" data-bs-backdrop="static" data-bs-keyboard="false" tabindex="-1">
-        <div class="modal-dialog">
-            <div class="modal-content">
-                <div class="modal-header">
-                    <h1 class="modal-title fs-5">Edit <?= isset($title) ? $title : '' ?></h1>
-                    <button type="button" class="btn-close" data-bs-dismiss="modal"></button>
-                </div>
-                <form id="form_${data.id}">
-                    <div class="modal-body">
-                        <div class="mb-3">
-                            <label for="judul" class="form-label">Judul</label>
-                            <input type="text" class="form-control" id="judul" name="judul" value="${data.judul}" placeholder="Masukkan judul">
-                            <div class="invalid-feedback" id="invalid_judul"></div>
-                        </div>
-                        <div class="mb-3">
-                            <label for="tautan" class="form-label">Tautan</label>
-                            <input type="text" class="form-control" id="tautan" name="tautan" value="${data.tautan}" placeholder="Masukkan tautan">
-                            <div class="invalid-feedback" id="invalid_tautan"></div>
-                        </div>
-                        <div class="mb-3">
-                            <label for="dokumen" class="form-label">Dokumen</label>
-                            <input type="file" class="form-control" id="dokumen" name="dokumen" accept="application/pdf">
-                            <div class="form-text">
-                                Maksimal 1 mb, pdf
-                            </div>
-                            <div class="invalid-feedback" id="invalid_dokumen"></div>
-                        </div>
-                    </div>
-                    <div class="modal-footer">
-                        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Batal</button>
-                        <button type="submit" class="btn btn-primary float-end">Simpan Perubahan</button>
-                    </div>
-                </form>
-            </div>
-        </div>
-    </div>
-    <a onclick="deleteData('${endpoint_hapus_data}')" title="Delete">
-        <i class="fa-regular fa-trash-can fa-lg text-danger"></i>
-    </a>`;
-
-    setTimeout(() => actionEdit(data.id), 0);
-
-    return html;
-}
-
-function actionEdit(id) {
-    const form = dom(`#form_${id}`);
-
-    if (! form.dataset.isInitialized) {
-        form.dataset.isInitialized = true;
-        form.addEventListener('submit', function(event) {
-            event.preventDefault();
-            const endpoint = `<?= $base_api ?>update/${id}`;
-            submitData(form, endpoint);
-        });
+    
+    if (data.created_by == '<?= userSession('id') ?>') {
+        return `
+        <a href="${endpoint_edit_data}" class="me-2" title="Edit">
+            <i class="fa-regular fa-pen-to-square fa-lg"></i>
+        </a>
+        <a onclick="deleteData('${endpoint_hapus_data}')" title="Delete">
+            <i class="fa-regular fa-trash-can fa-lg text-danger"></i>
+        </a>`;
+    } else {
+        return '';
     }
 }
 <?php endif; ?>
